+++ .dockerignore
@@ -1,0 +1,5 @@
+node_modules
+build
+.dockerignore
+Dockerfile
+Dockerfile.prod
