+++ jsconfig.json
@@ -1,0 +1,9 @@
+{
+    "compilerOptions": {
+      "baseUrl": "./src",
+      "paths": {
+        "components/*": ["components/*"]
+      }
+    },
+    "exclude": ["node_modules"] 
+  }
