+++ package.json
@@ -16,6 +16,7 @@
     "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
     "react-scroll": "^1.8.4",
+    "smooth-scrollbar": "^8.7.0",
     "three": "^0.134.0",
     "web-vitals": "^1.0.1"
   },
@@ -23,7 +24,9 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "docker:bash": "docker-compose -p bankaidevportfolio run --service-ports app"
+
   },
   "eslintConfig": {
     "extends": [
