+++ src/App.scss
@@ -14,5 +14,5 @@ $col-one: #fef5ed;
   transform: scale(1);
 }
 .App {
-  cursor: none;
+  // cursor: none;
 }
