--- src/js/Math.js
@@ -1,4 +1,0 @@
-/* eslint-disable no-extend-native */
-Number.prototype.map = function(in_min, in_max, out_min, out_max) {
-    return ((this - in_min) * (out_max - out_min)) / (in_max - in_min) + out_min
-  }
