+++ src/styles/_colors.scss
@@ -1,0 +1,26 @@
+$palette: (
+    // Primary
+    Black: #000000,
+    BaliHai: #82869A,
+    YellowOmlette: #ffa800,
+    DarkPuce:  #373C56,
+    VividOrange: #F0AE4B,
+    Red: #ff0101,
+    Facebookblue: #3a5897,
+    Royalblue: #4285f4,
+    BlueBase: #367BF5,
+    BlueSmooth: #E5EFFF,
+    Nero: #272727,
+    White: #ffffff,
+    Whisper: #e5e5e5,
+    Grey: #7a7a7a,
+    Grey1:#7D8EA3,
+    Grey5: #71717A,
+    Eclipse: #373737,
+    Whisper2: #eeeeee,
+    NavyBlue: #282f65,
+    NeutralGrey: #3e3e3e,
+    NeutralGrey2: #626262,
+    NeutralBlack: #242731,
+    PattensBlue: #E2E4E5
+);
