+++ src/styles/_setting.scss
@@ -1,0 +1,17 @@
+// Globals
+$global-font-size: 100%;
+// $global-font-family: "Nunito Sans", -apple-system, BlinkMacSystemFont,
+//   "Segoe UI", "Roboto", "Oxygen", "Ubuntu", "Cantarell", "Fira Sans",
+//   "Droid Sans", sans-serif;
+$header-font-family: "Roboto Mono", monospace;
+$text-font-family: "Work Sans", sans-serif;
+// Body styles
+$body-font-size: $global-font-size;
+$body-font-color: Black;
+
+// Font weights
+$font-weight-light: 300;
+$font-weight-regular: 400;
+$font-weight-dense: 500;
+$font-weight-medium: 600;
+$font-weight-bold: 700;
