+++ src/styles/global.scss
@@ -1,0 +1,2 @@
+//import Globals
+@import "globals/base";
