+++ src/styles/shared-styles.scss
@@ -1,0 +1,5 @@
+@import "utils/utils";
+@import "colors";
+@import "setting";
+@import "mixins/mixins";
+@import "global.scss";
